QCLI Automation for tools,model,editor,subscribe,usage,compact,hooks,issues,tests
